switch the malloc functions in kpathsea to (size_t) arguments to get rid of compiler warnings about casts from size_t to unsigned (texlive's kpathsea will have this same change soon)
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3553 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/kpathsea/lib.h
@@ -211,9 +211,9 @@ extern KPSEDLL off_t xftello (FILE *, const_string filename);
 /* These call the corresponding function in the standard library, and
    abort if those routines fail.  Also, `xrealloc' calls `xmalloc' if
    OLD_ADDRESS is null.  */
-extern KPSEDLL address xmalloc (unsigned size);
-extern KPSEDLL address xrealloc (address old_address, unsigned new_size);
-extern KPSEDLL address xcalloc (unsigned nelem, unsigned elsize);
+extern KPSEDLL address xmalloc (size_t size);
+extern KPSEDLL address xrealloc (address old_address, size_t new_size);
+extern KPSEDLL address xcalloc (unsigned nelem, size_t elsize);
 
 /* (Re)Allocate N items of type T using xmalloc/xrealloc.  */
 #define XTALLOC(n, t) ((t *) xmalloc ((n) * sizeof (t)))
